Добавляет слайс offers
